Refactored the nashorns away
Integrations are now handled on the Java side of things. This means
no more type definitions via javascript. This did not work out as planned
and would have required more and more awkward hacks. There was no easy
way to reobfuscate the javascript code on runtime or even during build time,
which meant that accessing most minecraft code did not work that way.

Shapes are not automatically being extracted anymore, but read from the JAR
directly if there is no file in the config folder with the same name.

This commit introduces a API to allow for easy integration by mods themselves,
but also integrates Forestry, Pams Harvestcraft etc itself using the same api.
The API is using annotated classes similiar to JEI. Documentation will follow.

This required some things to be changed:
- There is a new IntegrationRegistry used to find and track all the
  @BonsaiIntegration annotated classes.
- The TreeTypeRegistry is now being instanced and is available
  on the mod instance.
- Tree Drop chances are now floats between 0 and 1

TODO:
- It must be possible to remove integrations and modify individual trees,
  probably using crafttweaker
- API documentation