Added JSON defined tree types
Some integrations are so basic that its perfectly fine to define them
in a json file. Prime examples are IC2 and Integrated Dynamics, both
adding a single tree.

This also readds the ability for pack makers to define arbitrary custom
trees, e.g. a redstone tree grown from a piece of redstone dust.