Refactored grow time handling + CraftTweaker support
It should be possible to modify the individual grow times of the
registered trees. For that I needed to refactor a few things.
This commit adds a new ZenClass mods.bonsaitrees.TreeGrowth with
the method setTickMultiplier(<treename>, <multiplier>) to allow
setting the grow time of any tree supported by the mod.

Fixes #8