Added support for various soils.
Bonsai Pots now need soil for trees to grow on. Not every tree is
compatible with every soil and soils can affect the time it takes
for a tree to grow and its chances to drop items.

This should be a complete implementation of all the basic features,
which includes: CraftTweaker support, notification messages to help
the player, commands for modpack makers to debug and reload soils
and trees. It also includes a new Vanilla Chorus tree that can only
grow on End Stone and a few currently unused soil types. The readme
has been updated to reflect all the changes that have been made and
explains how soils can be configured.

Fixes #3