Added override to getMaxRenderDistanceSquared in TileBonsaiPot
This method is used to define the Tile's render distance and was before unused.
By using it, the Tile can stopped being rendered sooner than Minecraft's default 64 (?) blocks (my math might be wrong here) and make it, for example, dependent on the game's render distance.
This should overall increase performance.