Bumped to 2.0.2, aligned to latest libnonymous & fixes
- When listing saplings also print the compatible soils
  This made debugging of #138 much easier.
- Reduced log spam
- Fix bonsais not being able to be planted on servers
Some mod(s) trigger a reload early on when the server starts. Normally
that wouldn't be a problem, since we automatically rebuild the soil
compatibility list on the "RecipesUpdatedEvent". But that's only called
clientside, so we add this other way to also make sure it's being called
on the server.

Fixes #138