Added another command to auto-generate saplings and models
This should only be used on worlds specificially created for the purpose
of autogenerating sapling and model JSONs. It will use a custom super
flat dimension for this. It is WIP and might crash your game. It might
burn your world to the ground with it. You have been warned.