Allow specification of arbitrary upgrade items for drops
With this it is possible to restrict individual drops so they only
drop if a specific upgrade item is in one of the upgrade slots.
As all ingredients you can specify one or more items and tags that
would be valid. It is *not* possible to require two specific upgrades
at the same time.

This is still a bit experimental and the mod itself does not ship
with any recipes utilizing this (yet).

Fixes #257