Fix some loot drops not being generated: #371
The loot table for Chorus Flowers contains a `LootItemEntityPropertyCondition`
check that could never be fulfilled by a bonsai pot. We are now skipping the
conditions from (nested) loot tables that do not make sense for us.

This currently includes these four:
- ExplosionCondition
- LootItemBlockStatePropertyCondition
- LootItemEntityPropertyCondition
- LootItemKilledByPlayerCondition

It might probably make sense to turn this around - instead of blacklisting we
should whitelist all random chance and tool requirement conditions and ignore
the rest.